// app/generated/api/models/regionRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */
import type { RegionCode } from './regionCode';
import type { State } from './state';
import type { ZipCode } from './zipCode';
import type { BranchCode } from './branchCode';
import type { UpdatedBy } from './updatedBy';

export interface RegionRequest {
  regionCode: RegionCode;
  /**
   * @minLength 1
   * @maxLength 100
   */
  regionName: string;
  states: State[];
  zipCodes: ZipCode[];
  branches: BranchCode[];
  updatedBy: UpdatedBy;
}